Wrote DataHub module communication class with network support; Converted camera back to OpenCV capture; Disabled server; Moved server SID interpreter to AUV and Model message handlers; Regenerated Brain; Worked some more on the HAL;

+++ src/brain/HostLib_Audio.h
@@ -2,7 +2,7 @@
  * @file HostLib_Audio.h
  * @brief Helper for the AudioDevices.
  * Copyright 2007-2008 The MathWorks, Inc.
- * $Revision: 1.1.6.4.2.1 $ $Date: 2009/12/17 13:55:07 $
+ * $Revision: 1.1.6.6 $ $Date: 2010/02/17 18:47:46 $
  */ 
 
 
@@ -17,7 +17,7 @@ extern const char *libName_Audio;
  * Routines which are defined in the library in question
  *******************************/
 typedef void (*pFnLibCreate_Audio)(char *err, char *warn, const char *deviceName, int apiID, int inOut, void **m_device,
-                             int numChannels, double sampleRate, int deviceDatatype, int bufferSize, double queueDuration,
+                             int numChannels, double sampleRate, int deviceDatatype, int bufferSize, int queueDuration,
                              int frameSize, int signalDatatype);
 typedef void (*pFnLibUpdate_Audio)(void *device, char *err, const void *src, int signalDatatype, int samplesPerFrame, int *nDroppedSamples);
 typedef void (*pFnLibOutputs_Audio)(void *device, char *err, void *src, int signalDatatype, int samplesPerFrame, int *nDroppedSamples);
@@ -26,7 +26,7 @@ typedef void (*pFnLibOutputs_Audio)(void *device, char *err, void *src, int sign
  * Routines which we define to call the functions in the library 
  *******************************/
 void LibCreate_Audio(void *hostLib, char *warn, const char *deviceName, int apiID, int inOut,
-                     int numChannels, double sampleRate, int deviceDatatype, int bufferSize, double queueDuration,
+                     int numChannels, double sampleRate, int deviceDatatype, int bufferSize, int queueDuration,
                      int frameSize, int signalDatatype);
 void LibUpdate_Audio(void *hostLib, const void *src, int signalDatatype, int samplesPerFrame, int *nDroppedSamples);
 void LibOutputs_Audio(void *hostLib, void *src, int signalDatatype, int samplesPerFrame, int *nDroppedSamples);
